server/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt
pyjwt
cryptography